+++ lib/axlsx/workbook/worksheet/pivot_table.rb
@@ -26,6 +26,7 @@ module Axlsx
       @pages = []
       @subtotal = nil
       @no_subtotals_on_headers = []
+      @grand_totals = :both
       @sort_on_headers = {}
       @style_info = {}
       parse_options options
@@ -49,6 +50,19 @@ module Axlsx
       headers ||= {}
       headers = Hash[*headers.map { |h| [h, :ascending] }.flatten] if headers.is_a?(Array)
       @sort_on_headers = headers
+    end
+
+    # Defines which Grand Totals are to be shown.
+    # @return [Symbol] The row and/or column Grand Totals that are to be shown.
+    # Defaults to `:both` to show both row & column grand totals.
+    # Set to `:row_only`, `:col_only`, or `:none` to hide one or both Grand Totals.
+    attr_reader :grand_totals
+
+    # (see #grand_totals)
+    def grand_totals=(value)
+      RestrictionValidator.validate "PivotTable.grand_totals", [:both, :row_only, :col_only, :none], value
+
+      @grand_totals = value
     end
 
     # Style info for the pivot table
@@ -114,7 +128,8 @@ module Axlsx
       @columns = v
     end
 
-    # The data
+    # The data as an array of either headers (String) or hashes or mix of the two.
+    # Hash in format of { ref: header, num_fmt: numFmts, subtotal: subtotal }, where header is String, numFmts is Integer, and subtotal one of %w[sum count average max min product countNums stdDev stdDevp var varp]; leave subtotal blank to sum values
     # @return [Array]
     attr_reader :data
 
@@ -188,7 +203,11 @@ module Axlsx
     def to_xml_string(str = +'')
       str << '<?xml version="1.0" encoding="UTF-8"?>'
 
-      str << '<pivotTableDefinition xmlns="' << XML_NS << '" name="' << name << '" cacheId="' << cache_definition.cache_id.to_s << '"' << (data.size <= 1 ? ' dataOnRows="1"' : '') << ' applyNumberFormats="0" applyBorderFormats="0" applyFontFormats="0" applyPatternFormats="0" applyAlignmentFormats="0" applyWidthHeightFormats="1" dataCaption="Data" showMultipleLabel="0" showMemberPropertyTips="0" useAutoFormatting="1" indent="0" compact="0" compactData="0" gridDropZones="1" multipleFieldFilters="0">'
+      str << '<pivotTableDefinition xmlns="' << XML_NS << '" name="' << name << '" cacheId="' << cache_definition.cache_id.to_s << '"'
+      str << ' dataOnRows="1"' if data.size <= 1
+      str << ' rowGrandTotals="0"' if grand_totals == :col_only || grand_totals == :none
+      str << ' colGrandTotals="0"' if grand_totals == :row_only || grand_totals == :none
+      str << ' applyNumberFormats="0" applyBorderFormats="0" applyFontFormats="0" applyPatternFormats="0" applyAlignmentFormats="0" applyWidthHeightFormats="1" dataCaption="Data" showMultipleLabel="0" showMemberPropertyTips="0" useAutoFormatting="1" indent="0" compact="0" compactData="0" gridDropZones="1" multipleFieldFilters="0">'
 
       str << '<location firstDataCol="1" firstDataRow="1" firstHeaderRow="1" ref="' << ref << '"/>'
       str << '<pivotFields count="' << header_cells_count.to_s << '">'
@@ -244,8 +263,9 @@ module Axlsx
       unless data.empty?
         str << "<dataFields count=\"#{data.size}\">"
         data.each do |datum_value|
-          # The correct name prefix in ["Sum","Average", etc...]
-          str << "<dataField name='#{datum_value[:subtotal] || ''} of #{datum_value[:ref]}' fld='#{header_index_of(datum_value[:ref])}' baseField='0' baseItem='0'"
+          subtotal_name = datum_value[:subtotal] || 'sum'
+          subtotal_name = 'count' if name == 'countNums' # both count & countNums are labelled as count
+          str << "<dataField name='#{subtotal_name.capitalize} of #{datum_value[:ref]}' fld='#{header_index_of(datum_value[:ref])}' baseField='0' baseItem='0'"
           str << " numFmtId='#{datum_value[:num_fmt]}'" if datum_value[:num_fmt]
           str << " subtotal='#{datum_value[:subtotal]}' " if datum_value[:subtotal]
           str << "/>"
@@ -311,7 +331,11 @@ module Axlsx
       elsif columns.include? cell_ref
         attributes << 'axis="axisCol"'
         attributes << "sortType=\"#{sorttype == :descending ? 'descending' : 'ascending'}\"" if sorttype
-        include_items_tag = true
+        if subtotal
+          include_items_tag = true
+        else
+          attributes << 'defaultSubtotal="0"'
+        end
       elsif pages.include? cell_ref
         attributes << 'axis="axisPage"'
         include_items_tag = true